% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{d_legions}
\alias{d_legions}
\title{Dimensional data on the Legiones Astartes}
\format{
A tibble with 20 rows and the following fields:
\describe{
\item{id}{dbl The primary key for the legion}
\item{numeral}{chr The Roman numeral that identifies the legion}
\item{name}{chr The name of the legion}
\item{loyalty}{chr The legion's status during and after the Horus Heresy}
}
}
\usage{
d_legions
}
\description{
While technically only 30,000 years in the future, this is foundational
information for much of the subsequent 20,000.
}
\keyword{datasets}
