% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{d_planets}
\alias{d_planets}
\title{Dimensional data on planets}
\format{
A tibble with 20 rows and the following fields:
\describe{
\item{id}{dbl (Primary key) The primary key for the planet}
\item{is_special_code}{bool Denotes whether this is a 'special' code, such as Unknown or \link{REDACTED}. These are "sprinkled" throughout because, hey, there are a billion worlds, right? The Administratum deemed it inefficient to start numbering these from 999,999,999...}
\item{name}{chr The name by which the planet is best known.}
}
}
\usage{
d_planets
}
\description{
This table provides standard identification of planets for other tables to
connect to.
}
\keyword{datasets}
