% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{d_primarchs}
\alias{d_primarchs}
\title{Dimensional data on the Primarchs}
\format{
A tibble with 20 rows and the following fields:
\describe{
\item{id}{dbl (Primary key) The primary key for the Primarch}
\item{name}{chr The name of the Primarch}
\item{discovery_date}{chr The approximate date the Primarch was rediscovered}
\item{home_planet}{dbl (Foreign key) The ID of the planet on which the Primarch was rediscovered}
\item{legion}{dbl (Foreign key) The ID of the Primarch's legion}
}
}
\usage{
d_primarchs
}
\description{
While technically only 30,000 years in the future, this is foundational
information for much of the subsequent 20,000.
}
\keyword{datasets}
