% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{f_dates}
\alias{f_dates}
\title{Key dates and facts about them}
\format{
A tibble with 20 rows and the following fields:
\describe{
\item{date}{chr The approximate date of the observation}
\item{note}{chr Notable occurrences}
}
}
\usage{
f_dates
}
\description{
This provides key facts about interesting dates.
}
\keyword{datasets}
