% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{f_legions}
\alias{f_legions}
\title{Facts about the Legiones Astartes}
\format{
A tibble with 20 rows and the following fields:
\describe{
\item{legion}{dbl (Foreign key) The ID of the relevant legion}
\item{date}{chr The approximate date of the observation}
\item{strength}{dbl The approximate number of Space Marines in the legion}
}
}
\source{
https://warhammer40k.fandom.com/wiki/Space_Marine_Legions
}
\usage{
f_legions
}
\description{
This provides key facts about the legions.
}
\keyword{datasets}
